new version of the deployer plugin. Now it uses a declarative interface to describe the deployment instead of a programmatic one.
