+++ Package.swift
@@ -20,7 +20,7 @@ let package = Package(
         // plugin to deploy the lambda, relies on AWS SAM command line
         .plugin(name: "AWSLambdaDeployer", targets: ["AWSLambdaDeployer"]),
         // Shared Library to generate a SAM deployment descriptor
-        .library(name: "AWSLambdaDeploymentDescriptor", targets: ["AWSLambdaDeploymentDescriptor"]),
+        .library(name: "AWSLambdaDeploymentDescriptor", type: .dynamic, targets: ["AWSLambdaDeploymentDescriptor"]),
         // for testing only
         .library(name: "AWSLambdaTesting", targets: ["AWSLambdaTesting"]),
     ],
@@ -29,7 +29,6 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-log.git", .upToNextMajor(from: "1.4.2")),
         .package(url: "https://github.com/swift-server/swift-backtrace.git", .upToNextMajor(from: "1.2.3")),
         .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.0.0"),
-        .package(url: "https://github.com/jpsim/Yams.git", from: "5.0.1")
     ],
     targets: [
         .target(name: "AWSLambdaRuntime", dependencies: [
@@ -56,9 +55,6 @@ let package = Package(
         ),
         .target(
             name: "AWSLambdaDeploymentDescriptor",
-            dependencies: [
-                .product(name: "Yams", package: "Yams")
-            ],
             path: "Sources/AWSLambdaDeploymentDescriptor"
         ),
         .plugin(
